Incorporating filename into id

+++ shiftmedia/storage.py
@@ -4,6 +4,18 @@ from pathlib import Path
 from shiftmedia import exceptions as x
 from shiftmedia import utils
 
+# TODO: how to force file download?
+# TODO: how to serve original filename?
+# TODO: files for storage as is vs. file for processing
+
+# USE CASE: we might want file stored under its original name, i.e. cv.pdf
+
+# TODO: make original filename part of the ID?
+# TODO: 3c72aedc-ba25-11e6-a569-406c8f413974-original_filename.jpg
+# TODO: how do we transform such an id to original url?
+
+
+
 
 class Storage:
 
@@ -32,16 +44,21 @@ class Storage:
         # todo: validate file before put
         # todo: get real file extension to generate id with
 
+        # todo: whos responsibility it is to get file extension?
+        # todo: is it storage?
+
         if not os.path.exists(src):
             msg = 'Unable to find local file [{}]'
             raise x.LocalFileNotFound(msg.format(src))
 
-        extension = ''.join(Path(src).suffixes)[1:]
-        filename = 'original.' + extension
-        id = utils.generate_id(extension)
-        self.backend.put(src, id, filename)
-        if delete_local:
-            os.remove(src)
+        path = Path(src)
+        extension = ''.join(path.suffixes)[1:]
+        name = path.name.replace('.' + extension, '')
+        extension = utils.normalize_extension(extension)
+        filename = name + '.' + extension
+        id = utils.generate_id(filename)
+        self.backend.put(src, id)
+        if delete_local: os.remove(src)
         return id
 
     def delete(self, id):
